@@ -10,7 +10,7 @@
     "wiki": "http://public.kitware.com/Wiki/CDash"
   },
   "require": {
-    "php": "^7.1.3",
+    "php": "^7.3",
     "ext-bcmath": "*",
     "ext-curl": "*",
     "ext-json": "*",
@@ -26,17 +26,18 @@
     "friendsofphp/php-cs-fixer": "^2.16",
     "guzzlehttp/guzzle": "~6.2",
     "iron-io/iron_mq": "~2.0",
-    "knplabs/github-api": "dev-master#8c26dc85",
+    "knplabs/github-api": "^3.2",
     "laravel/framework": "6.20.14",
     "lcobucci/jwt": "3.2.0",
     "league/oauth2-github": "^2.0",
     "league/oauth2-google": "^2.0",
     "monolog/monolog": "~1.19",
+    "nyholm/psr7": "^1.4",
     "omines/oauth2-gitlab": "^3.1",
     "paragonie/random_compat": "2.0.4",
     "pear/archive_tar": "~1.4",
-    "php-di/php-di": "^6.1",
-    "php-http/guzzle6-adapter": "^1.1",
+    "php-di/php-di": "6.3.2",
+    "php-http/guzzle6-adapter": "2",
     "predis/predis": "~0.8",
     "ramsey/uuid": "~3.4",
     "sendgrid/sendgrid": "5.0.4"
